+++ src/lib.rs
@@ -163,8 +163,8 @@ pub use packet::{
     TsPacketIter, TsPacketLayout, MAX_ADAPTATION_FIELD_LEN, TS_PACKET_LEN, TS_SYNC_BYTE,
 };
 pub use pes::{
-    CoefficientSelection, DsmTrickMode, FieldId, PStdBuffer, PesExtension, PesPacket,
-    PesReassembler, ProgramPacketSequenceCounter,
+    CoefficientSelection, DsmTrickMode, FieldId, PStdBuffer, PesExtension, PesHeaderSpec,
+    PesPacket, PesReassembler, ProgramPacketSequenceCounter,
 };
 pub use psi::{
     decode_utc_time, iter_sections, mpeg2_crc32, BatTransportStream, BouquetAssociationTable,